Tag notifications
Tag notifications appear when tags are created, deleted, or updated

+++ app/Http/Controllers/TagController.php
@@ -3,6 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Http\Resources\TagResource;
+use App\Notifications\TagCreated;
+use App\Notifications\TagDeleted;
+use App\Notifications\TagUpdated;
 use Illuminate\Http\Request;
 use App\Models\Tag;
 use App\Http\Requests\StoreTagRequest;
@@ -26,8 +29,8 @@ class TagController extends Controller
     public function store(StoreTagRequest $request)
     {
         $validated = $request->validated();
-        Tag::create($validated);
-
+        $tag = Tag::create($validated);
+        $request->user()->notify(new TagCreated($tag));
         return response()->noContent(201);
     }
 
@@ -50,15 +53,21 @@ class TagController extends Controller
         $validated = $request->validated();
         
         $tag->update($validated);
+        $changes = $tag->getChanges();
+        $request->user()->notify(new TagUpdated($tag, $changes));
         return response('',200);
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Tag $tag)
+    public function destroy(Tag $tag, Request $request)
     {
+        if($request->user()->cannot('admin')){
+            abort(403);
+        }
         $tag->delete();
+        $request->user()->notify(new TagDeleted());
         return response()->noContent();
     }
 }
